move raredisease models to the analysis starter directory
