+++ cg/services/analysis_starter/configurator/file_creators/params_file/raredisease.py
@@ -6,12 +6,14 @@ from cg.constants.scout import ScoutExportFileName
 from cg.constants.tb import AnalysisType
 from cg.exc import CgDataError
 from cg.io.yaml import read_yaml, write_yaml_nextflow_style
-from cg.models.raredisease.raredisease import RarediseaseParameters
 from cg.services.analysis_starter.configurator.file_creators.params_file.abstract import (
     ParamsFileCreator,
 )
 from cg.services.analysis_starter.configurator.file_creators.params_file.utils import (
     replace_values_in_params_file,
+)
+from cg.services.analysis_starter.configurator.file_creators.sample_sheet.models import (
+    RarediseaseParameters,
 )
 from cg.store.models import BedVersion, Case, Sample
 from cg.store.store import Store
